requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
python-dotenv>=1.0.0
mysqlclient>=2.2.0
requests>=2.31
Pillow>=10.0.0